requirements.txt
aredis>=1.1.8,<1.2
discord.py>1.5,<1.7